+++ ethpm_types/ast.py
@@ -1,11 +1,7 @@
 from enum import Enum
 from typing import Dict, Iterator, List, Optional, Union
 
-try:
-    from pydantic.v1 import root_validator
-except ImportError:
-    from pydantic import root_validator
-
+from ethpm_types._pydantic_v1 import root_validator
 from ethpm_types.base import BaseModel
 from ethpm_types.sourcemap import SourceMapItem
 from ethpm_types.utils import SourceLocation
