+++ ethpm_types/source.py
@@ -4,11 +4,7 @@ from typing import Dict, Iterator, List, Optional, Set, Tuple, Union
 import requests
 from cid import make_cid  # type: ignore
 
-try:
-    from pydantic.v1 import root_validator, validator
-except ImportError:
-    from pydantic import root_validator, validator
-
+from ethpm_types._pydantic_v1 import root_validator, validator
 from ethpm_types.ast import ASTClassification, ASTNode, SourceLocation
 from ethpm_types.base import BaseModel
 from ethpm_types.contract_type import ContractType
