+++ pyproject.toml
@@ -35,5 +35,5 @@ multi_line_output = 3
 use_parentheses = true
 
 [tool.mypy]
-exclude = "build"
+exclude = ["build", "ethpm_types/_pydantic_v1.py"]
 plugins = ["pydantic.mypy"]
