+++ setup.cfg
@@ -5,3 +5,4 @@ exclude =
 	.eggs
 	docs
 	build
+	_pydantic_v1.py
