+++ tests/test_package_manifest.py
@@ -4,12 +4,8 @@ import github
 import pytest
 import requests
 
-try:
-    from pydantic.v1 import ValidationError
-except ImportError:
-    from pydantic import ValidationError
-
-from ethpm_types import PackageManifest
+from ethpm_types._pydantic_v1 import ValidationError
+from ethpm_types.manifest import PackageManifest
 
 ETHPM_SPEC_REPO = github.Github(os.environ.get("GITHUB_ACCESS_TOKEN", None)).get_repo(
     "ethpm/ethpm-spec"
